+++ src/api/objects/toot.ts
@@ -285,8 +285,12 @@ export default class Toot implements TootObj {
         repairAccount(this.account);
         this.mentions.forEach(repairAccount);
 
-        if (this.reblog?.account && !this.reblogsByAccts().includes(this.account.acct)) {
-            this.reblog.reblogsBy.push(this.account);
+        if (this.reblog){
+            this.trendingRank ||= this.reblog.trendingRank;
+
+            if (!this.reblogsByAccts().includes(this.account.acct)) {
+                this.reblog.reblogsBy.push(this.account);
+            }
         }
 
         // Check for weird media types
@@ -310,7 +314,7 @@ export default class Toot implements TootObj {
         const prefix = logLabel ? `[${logLabel}] ` : '';
         const tootsByURI = groupBy<Toot>(toots, toot => toot.realURI());
 
-        Object.entries(tootsByURI).forEach(([_uri, uriToots]) => {
+        Object.values(tootsByURI).forEach((uriToots) => {
             const allTrendingTags = uriToots.flatMap(toot => toot.trendingTags || []);
             const uniqueTrendingTags = [...new Map(allTrendingTags.map((tag) => [tag.name, tag])).values()];
             const firstScoredToot = uriToots.find(toot => !!toot.scoreInfo);
@@ -326,6 +330,7 @@ export default class Toot implements TootObj {
                 // Set missing scoreInfo to first scoreInfo we can find (if any)
                 toot.scoreInfo ??= firstScoredToot?.scoreInfo;
                 toot.trendingRank ??= firstRankedToot?.trendingRank;
+                if (toot.reblog) toot.reblog.trendingRank ??= firstRankedToot?.trendingRank;
                 // Set reblogsBy
                 toot.reblogsBy = reblogsBy;
             });